@@ -2,9 +2,29 @@
 
 namespace Laraditz\Shopee\Enums;
 
-class ShopStatus extends Enums
+use Exception;
+
+enum ShopStatus: int
 {
-    const Normal    = 1;
-    const Banned    = 2;
-    const Frozen    = 3;
+    case Normal = 1;
+    case Banned = 2;
+    case Frozen = 3;
+
+    public static function fromName(string $name)
+    {
+        try {
+            return self::{$name};
+        } catch (\Throwable $th) {
+            throw new Exception("$name is not a valid backing value for enum " . self::class);
+        }
+    }
+
+    public static function tryFromName(string $name)
+    {
+        try {
+            return self::fromName($name);
+        } catch (\Throwable $th) {
+            return null;
+        }
+    }
 }
