@@ -2,8 +2,9 @@
 
 namespace Laraditz\Shopee\Services;
 
-use Laraditz\Shopee\Models\ShopeeOrder;
 use Laraditz\Shopee\Models\ShopeeShop;
+use Laraditz\Shopee\Models\ShopeeOrder;
+use Laraditz\Shopee\Models\ShopeeRequest;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class OrderService extends BaseService
@@ -50,27 +51,34 @@ class OrderService extends BaseService
         return null;
     }
 
-    public function detail(string $order_sn, array $extraFields = [])
+    public function detail(string $order_sn, array $extra_fields = [], ?int $shop_id = null)
     {
-        $order = ShopeeOrder::findOrFail($order_sn);
+        $order = ShopeeOrder::find($order_sn);
+        $shop = null;
+
+        if (!$order && $shop_id) {
+            $shop = ShopeeShop::findOrFail($shop_id);
+        } elseif ($order) {
+            $shop = $order?->shop;
+        }
+
         throw_if(
-            !$order->shop,
+            !$shop,
             NotFoundHttpException::class,
             'Shop not found.'
         );
-        $shop = $order->shop;
 
         $partner_id = app('shopee')->getPartnerId();
         $route = 'order.get_detail';
         $path = app('shopee')->getPath($route);
         $access_token = data_get($shop, 'accessToken.access_token');
-        $signature = app('shopee')->helper()->generateSignature($path, [$access_token, $order->shop_id]);
+        $signature = app('shopee')->helper()->generateSignature($path, [$access_token, $shop->id]);
 
         $query_string = [
             'partner_id' => $partner_id,
             'timestamp' => $signature['time'],
             'access_token' => $access_token,
-            'shop_id' => $order->shop_id,
+            'shop_id' => $shop->id,
             'sign' => $signature['signature'],
         ];
 
@@ -78,8 +86,8 @@ class OrderService extends BaseService
             'order_sn_list' => $order_sn,
         ];
 
-        if (count($extraFields) > 0) {
-            $payload['response_optional_fields'] = implode(',', $extraFields);
+        if (count($extra_fields) > 0) {
+            $payload['response_optional_fields'] = implode(',', $extra_fields);
         }
 
         $response = $this->route($route)
@@ -92,5 +100,35 @@ class OrderService extends BaseService
         }
 
         return null;
+    }
+
+    public function afterGetOrderListResponse(ShopeeRequest $request, ?array $result = [])
+    {
+        $order_list = data_get($result, 'response.order_list');
+
+        if ($order_list && count($order_list) > 0) {
+            foreach ($order_list as $order) {
+                ShopeeOrder::firstOrCreate([
+                    'id' => $order['order_sn'],
+                    'shop_id' => $request->shop_id
+                ]);
+            }
+        }
+    }
+
+    public function afterGetOrderDetailResponse(ShopeeRequest $request, ?array $result = [])
+    {
+        $order_list = data_get($result, 'response.order_list');
+
+        if ($order_list && count($order_list) > 0) {
+            foreach ($order_list as $order) {
+                ShopeeOrder::updateOrCreate([
+                    'id' => $order['order_sn'],
+                    'shop_id' => $request->shop_id
+                ], [
+                    'status' => $order['order_status'] ?? null,
+                ]);
+            }
+        }
     }
 }
